requirements.txt
numpy
pandas
matplotlib
opencv-python
torch
torchvision
tqdm
pyyaml
scikit-learn
